recommendations/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
pytest ~= 5.4
flask ~= 1.1
Jinja2 ~= 2.11
numpy
mysql-connector-python